pyproject.toml
[project]
name = "microblog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiosmtpd>=1.4.6",
    "alembic==1.17.2",
    "blinker==1.9.0",
    "click==8.3.1",
    "email-validator>=2.3.0",
    "flask==3.1.2",
    "flask-babel>=4.0.0",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-migrate==4.1.0",
    "flask-moment>=1.0.6",
    "flask-sqlalchemy==3.1.1",
    "flask-wtf==1.2.2",
    "greenlet==3.3.0",
    "itsdangerous==2.2.0",
    "jinja2==3.1.6",
    "langdetect>=1.0.9",
    "mako==1.3.10",
    "markupsafe==3.0.3",
    "pyjwt>=2.10.1",
    "python-dotenv==1.2.1",
    "python-iso639>=2025.11.16",
    "requests>=2.32.5",
    "sqlalchemy==2.0.45",
    "tomli==2.3.0",
    "typing-extensions==4.15.0",
    "werkzeug==3.1.4",
    "wtforms==3.2.1",
]
